pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asset-forge-mcp"
version = "0.1.0"
description = "MCP server for generating and editing game assets via OpenAI Image API"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.8.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
    "uvicorn>=0.30",
]

[project.scripts]
asset-forge-mcp = "asset_forge_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/asset_forge_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
